Fix Deletion Tests
